Refactoring complete
