+++ cmake-build-debug/.cmake/api/v1/reply/target-CourseWork-Debug-858e6061fbcdfb916b3e.json
@@ -63,7 +63,10 @@
 				9,
 				11,
 				13,
-				15
+				15,
+				17,
+				19,
+				21
 			]
 		}
 	],
@@ -108,7 +111,10 @@
 				9,
 				11,
 				13,
-				15
+				15,
+				17,
+				19,
+				21
 			]
 		},
 		{
@@ -122,7 +128,10 @@
 				10,
 				12,
 				14,
-				16
+				16,
+				18,
+				20,
+				22
 			]
 		}
 	],
@@ -159,12 +168,12 @@
 		{
 			"backtrace" : 1,
 			"compileGroupIndex" : 0,
-			"path" : "Tree.cpp",
+			"path" : "archiver/Tree.cpp",
 			"sourceGroupIndex" : 0
 		},
 		{
 			"backtrace" : 1,
-			"path" : "Tree.h",
+			"path" : "archiver/Tree.h",
 			"sourceGroupIndex" : 1
 		},
 		{
@@ -220,6 +229,39 @@
 		{
 			"backtrace" : 1,
 			"path" : "utils/Utils.h",
+			"sourceGroupIndex" : 1
+		},
+		{
+			"backtrace" : 1,
+			"compileGroupIndex" : 0,
+			"path" : "archiver/BodyBinary.cpp",
+			"sourceGroupIndex" : 0
+		},
+		{
+			"backtrace" : 1,
+			"path" : "archiver/BodyBinary.h",
+			"sourceGroupIndex" : 1
+		},
+		{
+			"backtrace" : 1,
+			"compileGroupIndex" : 0,
+			"path" : "archiver/FanoCodesBinary.cpp",
+			"sourceGroupIndex" : 0
+		},
+		{
+			"backtrace" : 1,
+			"path" : "archiver/FanoCodesBinary.h",
+			"sourceGroupIndex" : 1
+		},
+		{
+			"backtrace" : 1,
+			"compileGroupIndex" : 0,
+			"path" : "binary/BinaryIO.cpp",
+			"sourceGroupIndex" : 0
+		},
+		{
+			"backtrace" : 1,
+			"path" : "binary/BinaryIO.h",
 			"sourceGroupIndex" : 1
 		}
 	],
