+++ cmake-build-debug/CMakeFiles/CourseWork.dir/ArchiverExecutor.cpp.obj.d
@@ -151,7 +151,7 @@ CMakeFiles/CourseWork.dir/ArchiverExecutor.cpp.obj: \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/istream \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/istream.tcc \
  C:\Users\NHTHS\Desktop\CourseWork\Fano.h \
- C:\Users\NHTHS\Desktop\CourseWork\Tree.h \
+ C:\Users\NHTHS\Desktop\CourseWork\archiver/Tree.h \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/vector \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/stl_uninitialized.h \
  C:/PROGRA~1/JETBRA~1/CLION2~1.2/bin/mingw/lib/gcc/x86_64-w64-mingw32/11.2.0/include/c++/bits/stl_vector.h \
