+++ cmake-build-debug/CMakeFiles/CourseWork.dir/DependInfo.cmake
@@ -11,8 +11,11 @@ set(CMAKE_DEPENDS_DEPENDENCY_FILES
   "C:/Users/NHTHS/Desktop/CourseWork/Archive.cpp" "CMakeFiles/CourseWork.dir/Archive.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/Archive.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/ArchiverExecutor.cpp" "CMakeFiles/CourseWork.dir/ArchiverExecutor.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/ArchiverExecutor.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/Fano.cpp" "CMakeFiles/CourseWork.dir/Fano.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/Fano.cpp.obj.d"
+  "C:/Users/NHTHS/Desktop/CourseWork/archiver/BodyBinary.cpp" "CMakeFiles/CourseWork.dir/archiver/BodyBinary.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/archiver/BodyBinary.cpp.obj.d"
+  "C:/Users/NHTHS/Desktop/CourseWork/archiver/FanoCodesBinary.cpp" "CMakeFiles/CourseWork.dir/archiver/FanoCodesBinary.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/archiver/FanoCodesBinary.cpp.obj.d"
+  "C:/Users/NHTHS/Desktop/CourseWork/archiver/Tree.cpp" "CMakeFiles/CourseWork.dir/archiver/Tree.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/archiver/Tree.cpp.obj.d"
+  "C:/Users/NHTHS/Desktop/CourseWork/binary/BinaryIO.cpp" "CMakeFiles/CourseWork.dir/binary/BinaryIO.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/binary/BinaryIO.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/main.cpp" "CMakeFiles/CourseWork.dir/main.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/main.cpp.obj.d"
-  "C:/Users/NHTHS/Desktop/CourseWork/Tree.cpp" "CMakeFiles/CourseWork.dir/tree.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/tree.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/utils/ConfigMapper.cpp" "CMakeFiles/CourseWork.dir/utils/ConfigMapper.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/utils/ConfigMapper.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/utils/Exception.cpp" "CMakeFiles/CourseWork.dir/utils/Exception.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/utils/Exception.cpp.obj.d"
   "C:/Users/NHTHS/Desktop/CourseWork/utils/Log.cpp" "CMakeFiles/CourseWork.dir/utils/Log.cpp.obj" "gcc" "CMakeFiles/CourseWork.dir/utils/Log.cpp.obj.d"
