+++ cmake-build-debug/CMakeFiles/CourseWork.dir/cmake_clean.cmake
@@ -5,8 +5,14 @@ file(REMOVE_RECURSE
   "CMakeFiles/CourseWork.dir/ArchiverExecutor.cpp.obj.d"
   "CMakeFiles/CourseWork.dir/Fano.cpp.obj"
   "CMakeFiles/CourseWork.dir/Fano.cpp.obj.d"
-  "CMakeFiles/CourseWork.dir/Tree.cpp.obj"
-  "CMakeFiles/CourseWork.dir/Tree.cpp.obj.d"
+  "CMakeFiles/CourseWork.dir/archiver/BodyBinary.cpp.obj"
+  "CMakeFiles/CourseWork.dir/archiver/BodyBinary.cpp.obj.d"
+  "CMakeFiles/CourseWork.dir/archiver/FanoCodesBinary.cpp.obj"
+  "CMakeFiles/CourseWork.dir/archiver/FanoCodesBinary.cpp.obj.d"
+  "CMakeFiles/CourseWork.dir/archiver/Tree.cpp.obj"
+  "CMakeFiles/CourseWork.dir/archiver/Tree.cpp.obj.d"
+  "CMakeFiles/CourseWork.dir/binary/BinaryIO.cpp.obj"
+  "CMakeFiles/CourseWork.dir/binary/BinaryIO.cpp.obj.d"
   "CMakeFiles/CourseWork.dir/main.cpp.obj"
   "CMakeFiles/CourseWork.dir/main.cpp.obj.d"
   "CMakeFiles/CourseWork.dir/utils/ConfigMapper.cpp.obj"
