+++ cmake-build-debug/CMakeFiles/CourseWork.dir/progress.make
@@ -8,4 +8,7 @@ CMAKE_PROGRESS_7 = 7
 CMAKE_PROGRESS_8 = 8
 CMAKE_PROGRESS_9 = 9
 CMAKE_PROGRESS_10 = 10
+CMAKE_PROGRESS_11 = 11
+CMAKE_PROGRESS_12 = 12
+CMAKE_PROGRESS_13 = 13
