+++ cmake-build-debug/Testing/Temporary/LastTest.log
@@ -1,3 +1,3 @@
-Start testing: Nov 30 18:28 RTZ 2 (ceia)
+Start testing: Nov 30 19:25 RTZ 2 (ceia)
 ----------------------------------------------------------
-End testing: Nov 30 18:28 RTZ 2 (ceia)
+End testing: Nov 30 19:25 RTZ 2 (ceia)
